feat: complete tiny-skia migration with window and input support
- Fix event loop handling in simulator display initialization
- Add event_loop.exit() call after window creation to prevent blocking
- Add bounds checking for background image loading
- Add debug logging for window and key event debugging
- Fix ESC key mapping to Menu key in test example
- Create comprehensive test_rendering.rs example

All rendering now works with tiny-skia and the simulator window
displays correctly with functional keyboard input.

+++ crates/allium-menu/src/view/guide_selector.rs
@@ -72,7 +72,7 @@ impl GuideSelector {
 
         let button_hints_rect = button_hints.bounding_box(&styles);
         let list_rect = Rect::new(
-            rect.x + styles.ui.margin_x as i32,
+            rect.x + styles.ui.margin_x,
             rect.y,
             rect.w - (styles.ui.margin_x * 2) as u32,
             button_hints_rect.y as u32 - rect.y as u32,
